+++ src/s2geography/geoarrow-imports.h
@@ -49,7 +49,7 @@ class Handler {
   virtual Result geom_start(util::GeometryType geometry_type, int64_t size) {
     return Result::CONTINUE;
   }
-  virtual Result ring_start(int32_t size) { return Result::CONTINUE; }
+  virtual Result ring_start(int64_t size) { return Result::CONTINUE; }
   virtual Result coords(const double* coord, int64_t n, int32_t coord_size) {
     return Result::CONTINUE;
   }
